+++ shv/rpcserver.py
@@ -1,15 +1,15 @@
 """RPC server that waits for clients connection."""
+from __future__ import annotations
+
 import abc
 import asyncio
 import collections.abc
-import contextlib
-import functools
 import logging
 import typing
 
-from .chainpack import ChainPack
-from .rpcclient import RpcClient, RpcClientStream
-from .rpcmessage import RpcMessage
+from . import rpcprotocol
+from .rpcclient import RpcClient
+from .rpcprotocol import RpcProtocolSerial, RpcProtocolStream, RpcTransportProtocol
 from .rpcurl import RpcProtocol, RpcUrl
 
 logger = logging.getLogger(__name__)
@@ -19,236 +19,202 @@ class RpcServer(abc.ABC):
     """RPC server listening for new SHV connections."""
 
     @abc.abstractmethod
+    def is_serving(self) -> bool:
+        """Check if server is accepting new SHV connections."""
+
+    @abc.abstractmethod
+    async def listen(self) -> None:
+        """Start accepting new SHV connections."""
+
+    @abc.abstractmethod
+    async def listen_forewer(self) -> None:
+        """Listen and block the calling coroutine until cancelation."""
+
+    @abc.abstractmethod
     def close(self) -> None:
         """Stop accepting new SHV connections."""
 
+    @abc.abstractmethod
     async def wait_closed(self) -> None:
-        """Wait for close to be actually performed."""
+        """Stop accepting new SHV connections and for that to make effect."""
 
 
-class RpcServerStream(RpcServer):
-    """RPC server listening for new SHV connections on stream transport layer."""
+class _RpcServerStream(RpcServer):
+    """RPC server listenting for SHV connections for streams."""
 
     def __init__(
         self,
         client_connected_cb: typing.Callable[
             [RpcClient], None | collections.abc.Awaitable[None]
         ],
-        loop: asyncio.events.AbstractEventLoop,
-        get_server: typing.Callable,  # TODO
-        wait_listening: asyncio.Future,
+        protocol_factory: typing.Type[RpcTransportProtocol] = RpcProtocolStream,
     ):
-        """Do not initialize this object directly, use :meth:`tcp_listen` or
-        :meth:`unix_listen` instead.
-        """
-        self._client_connected_cb = client_connected_cb
+        self.client_connected_cb = client_connected_cb
+        """Callbact that is called when new client is connected."""
+        self.protocol_factory = protocol_factory
+        """Protocol factory used to create protocol for new clients."""
+        self.clients: list[_RpcServerStream.Client] = []
+        """List of clients used for termination of the server."""
         self._server: asyncio.Server | None = None
-        self._task = loop.create_task(self._serve(get_server, wait_listening))
 
-    async def _serve(
-        self,
-        get_server: typing.Callable,
-        wait_listening: asyncio.Future,
-    ) -> None:
-        self._server = await get_server(self._client_connect)
-        assert self._server is not None
-        wait_listening.set_result(self._server)
+    @abc.abstractmethod
+    async def _create_server(self) -> asyncio.Server:
+        """Create the server instance."""
+
+    def is_serving(self) -> bool:
+        return self._server is not None and self._server.is_serving()
+
+    async def listen(self) -> None:
+        if self._server is None:
+            self._server = await self._create_server()
+        await self._server.start_serving()
+
+    async def listen_forewer(self) -> None:
+        if self._server is None:
+            self._server = await self._create_server()
         await self._server.serve_forever()
 
     async def _client_connect(
         self, reader: asyncio.StreamReader, writer: asyncio.StreamWriter
     ) -> None:
-        client = RpcClientStream(reader, writer)
-        res = self._client_connected_cb(client)
+        client = self.Client(reader, writer, self)
+        self.clients.append(client)
+        res = self.client_connected_cb(client)
         if isinstance(res, collections.abc.Awaitable):
             await res
 
-    @classmethod
-    async def listen(
-        cls,
-        client_connected_cb: typing.Callable[
-            [RpcClient], None | collections.abc.Awaitable[None]
-        ],
-        location: str,
-        port: int,
-    ) -> "RpcServerStream":
-        """Start listening on given address and port for TCP connections.
-
-        :param client_connected_cb: Function called for every new connected client.
-        :param location: IP/Hostname
-        :param port: Listening port
-        :return: Handle used to control the listening server.
-        """
-        loop = asyncio.get_running_loop()
-        future = loop.create_future()
-        res = cls(
-            client_connected_cb,
-            loop,
-            functools.partial(asyncio.start_server, host=location, port=port),
-            future,
-        )
-        await future
-        return res
-
-    @classmethod
-    async def unix_listen(
-        cls,
-        client_connected_cb: typing.Callable[
-            [RpcClient], None | collections.abc.Awaitable[None]
-        ],
-        path: str,
-    ) -> "RpcServerStream":
-        """Start listening on given address and port for local socket connections.
-
-        :param client_connected_cb: Function called for every new connected client.
-        :param path: Path to the Unix domain socket.
-        :return: Handle used to control the listening server.
-        """
-        # TODO try to remove socket on server stop
-        loop = asyncio.get_running_loop()
-        future = loop.create_future()
-        res = cls(
-            client_connected_cb,
-            loop,
-            functools.partial(asyncio.start_unix_server, path=path),
-            future,
-        )
-        await future
-        return res
-
     def close(self) -> None:
-        assert self._server is not None
-        self._server.close()
+        if self._server is not None:
+            self._server.close()
 
     async def wait_closed(self) -> None:
-        assert self._server is not None
-        await self._server.wait_closed()
-        with contextlib.suppress(asyncio.CancelledError):
-            await self._task
+        if self._server is not None:
+            self.close()
+            await self._server.wait_closed()
 
+    class Client(RpcClient):
+        """RPC client for TCP/IP server connection."""
 
-class RpcServerDatagram(RpcServer):
-    """RPC server listening for new SHV connections on datagram transport layer."""
-
-    class _RpcClientDatagramServer(RpcClient):
         def __init__(
             self,
-            addr: tuple[str | typing.Any, int],
-            protocol: "RpcServerDatagram._Protocol",
+            reader: asyncio.StreamReader,
+            writer: asyncio.StreamWriter,
+            server: _RpcServerStream,
         ) -> None:
             super().__init__()
-            self._addr = addr
-            self._protocol = protocol
-
-        async def send(self, msg: RpcMessage) -> None:
-            await super().send(msg)
-            self._protocol.transport.sendto(
-                bytes((ChainPack.ProtocolType,)) + msg.to_chainpack(), self._addr
+            self._reader = reader
+            self._writer = writer
+            self._protocol = rpcprotocol.protocol_for_asyncio_stream(
+                server.protocol_factory, reader, writer
             )
 
-        async def _receive(self) -> bytes | None:
-            queue = self._protocol.existing_clients.get(self._addr, None)
-            if queue is None:
-                return None
-            data = await queue.get()
-            queue.task_done()
-            return data
+        async def _send(self, msg: bytes) -> None:
+            await self._protocol.send(msg)
 
-        async def _reset(self) -> None:
-            self._protocol.transport._sock.sendto(bytes(), self._addr)  # type: ignore
+        async def _receive(self) -> bytes:
+            return await self._protocol.receive()
 
+        @property
         def connected(self) -> bool:
-            return self._addr in self._protocol.existing_clients
+            return not self._writer.is_closing()
 
-        async def disconnect(self) -> None:
-            if self._addr in self._protocol.existing_clients:
-                await self._protocol.existing_clients[self._addr].put(None)
-            self._protocol.existing_clients.pop(self._addr, None)
-            if not self._protocol.existing_clients and not self._protocol.accept_new:
-                self._protocol.transport.close()
+        async def reset(self) -> bool:
+            self.disconnect()
+            return False
 
-    class _Protocol(asyncio.DatagramProtocol):
-        """Implementation of Asyncio datagram protocol to listen on UDP."""
+        def disconnect(self) -> None:
+            self._writer.close()
 
-        def __init__(
-            self,
-            new_client_cb: typing.Callable[
-                [RpcClient], None | collections.abc.Awaitable[None]
-            ],
-        ):
-            self.new_client_cb = new_client_cb
-            self.existing_clients: dict[
-                tuple[str | typing.Any, int], asyncio.Queue[bytes | None]
-            ] = {}
-            self.transport: asyncio.DatagramTransport
-            self.accept_new: bool = True
-            self._loop = asyncio.get_running_loop()
+        async def wait_disconnect(self) -> None:
+            self.disconnect()
+            await self._writer.wait_closed()
 
-        def connection_made(self, transport: asyncio.BaseTransport) -> None:
-            self.transport = typing.cast(asyncio.DatagramTransport, transport)
 
-        def datagram_received(
-            self, data: bytes, addr: tuple[str | typing.Any, int]
-        ) -> None:
-            if not data or addr not in self.existing_clients:
-                if not self.accept_new:
-                    return
-                self.existing_clients[addr] = asyncio.Queue()
-                res = self.new_client_cb(
-                    RpcServerDatagram._RpcClientDatagramServer(addr, self)
-                )
-                if asyncio.iscoroutine(res):
-                    self._loop.create_task(res)
-            if data:
-                self.existing_clients[addr].put_nowait(data)
+class RpcServerTCP(_RpcServerStream):
+    """RPC server listenting for SHV connections in TCP/IP."""
 
     def __init__(
         self,
-        transport: asyncio.DatagramTransport,
-        protocol: _Protocol,
-    ):
-        """Do not initialize this directly, use :func:`create_rpc_server` instead."""
-        self._transport = transport
-        self._protocol = protocol
-        self._close = asyncio.get_running_loop().create_future()
-
-    @classmethod
-    async def listen(
-        cls,
         client_connected_cb: typing.Callable[
             [RpcClient], None | collections.abc.Awaitable[None]
         ],
-        location: str,
-        port: int,
-    ) -> "RpcServerDatagram":
-        """Start listening on given address and port for TCP connections.
+        location: str | None = None,
+        port: int = 3755,
+        protocol_factory: typing.Type[RpcTransportProtocol] = RpcProtocolStream,
+    ):
+        super().__init__(client_connected_cb, protocol_factory)
+        self.location = location
+        self.port = port
 
-        :param client_connected_cb: Function called for every new connected client.
-        :param location: IP/Hostname
-        :param port: Listening port
-        :return: Handle used to control the listening server.
-        """
-        t, p = await asyncio.get_running_loop().create_datagram_endpoint(
-            lambda: cls._Protocol(client_connected_cb),
-            local_addr=(location or "localhost", port),
+    async def _create_server(self) -> asyncio.Server:
+        return await asyncio.start_server(
+            self._client_connect, host=self.location, port=self.port
         )
-        return cls(t, p)
+
+    async def listen(self) -> None:
+        await super().listen()
+        assert self._server is not None
+        if not self._server.is_serving():
+            logger.debug("Listening for clients: (TCP) %s:%d", self.location, self.port)
 
     def close(self) -> None:
-        """Stop accepting new SHV connections.
+        super().close()
+        if self._server is not None and self._server.is_serving():
+            logger.debug(
+                "No longer listening for clients: (TCP) %s:%d", self.location, self.port
+            )
 
-        Be aware that compared with TCP/IP this won't immediately free the port for
-        listening. The listening socket is actually closed only when server stops
-        listening for a new connections and all clients are disconnected.
-        """
-        self._protocol.accept_new = False
-        if not self._protocol.existing_clients:
-            self._transport.close()
-        if not self._close.done():
-            self._close.set_result(None)
+    async def _client_connect(
+        self, reader: asyncio.StreamReader, writer: asyncio.StreamWriter
+    ) -> None:
+        peername = writer.get_extra_info("peername")
+        logger.debug(
+            "New client: (TCP) %s:%d: %s:%d",
+            self.location,
+            self.port,
+            peername[0],
+            peername[1],
+        )
+        await super()._client_connect(reader, writer)
 
-    async def wait_closed(self) -> None:
-        await self._close
+
+class RpcServerUnix(_RpcServerStream):
+    """RPC server listenting for SHV connections on Unix domain named socket."""
+
+    def __init__(
+        self,
+        client_connected_cb: typing.Callable[
+            [RpcClient], None | collections.abc.Awaitable[None]
+        ],
+        location: str = "shv.sock",
+        protocol_factory: typing.Type[RpcTransportProtocol] = RpcProtocolStream,
+    ):
+        super().__init__(client_connected_cb, protocol_factory)
+        self.location = location
+
+    async def _create_server(self) -> asyncio.Server:
+        return await asyncio.start_unix_server(self._client_connect, path=self.location)
+
+    async def listen(self) -> None:
+        await super().listen()
+        assert self._server is not None
+        if not self._server.is_serving():
+            logger.debug("Listening for clients: (Unix) %s", self.location)
+
+    def close(self) -> None:
+        super().close()
+        if self._server is not None and self._server.is_serving():
+            logger.debug("No longer listening for clients: (Unix) %s", self.location)
+
+    async def _client_connect(
+        self, reader: asyncio.StreamReader, writer: asyncio.StreamWriter
+    ) -> None:
+        logger.debug(
+            "New client: (Unix) %s: %s",
+            self.location,
+            writer.get_extra_info("peername"),
+        )
+        await super()._client_connect(reader, writer)
 
 
 async def create_rpc_server(
@@ -262,12 +228,20 @@ async def create_rpc_server(
     :param client_connected_cb: function called for every new client connected.
     :param url: RPC URL specifying where server should listen.
     """
+    res: RpcServer
     if url.protocol is RpcProtocol.TCP:
-        return await RpcServerStream.listen(client_connected_cb, url.location, url.port)
-    if url.protocol is RpcProtocol.LOCAL_SOCKET:
-        return await RpcServerStream.unix_listen(client_connected_cb, url.location)
-    if url.protocol is RpcProtocol.UDP:
-        return await RpcServerDatagram.listen(
-            client_connected_cb, url.location, url.port
+        res = RpcServerTCP(
+            client_connected_cb, url.location, url.port, RpcProtocolStream
         )
-    raise NotImplementedError(f"Unimplemented protocol: {url.protocol}")
+    elif url.protocol is RpcProtocol.TCPS:
+        res = RpcServerTCP(
+            client_connected_cb, url.location, url.port, RpcProtocolSerial
+        )
+    elif url.protocol is RpcProtocol.UNIX:
+        res = RpcServerUnix(client_connected_cb, url.location, RpcProtocolStream)
+    elif url.protocol is RpcProtocol.UNIXS:
+        res = RpcServerUnix(client_connected_cb, url.location, RpcProtocolSerial)
+    else:
+        raise NotImplementedError(f"Unimplemented protocol: {url.protocol}")
+    await res.listen()
+    return res