+++ tests/conftest.py
@@ -1,4 +1,3 @@
-import asyncio
 import dataclasses
 import pathlib
 import socket