+++ tests/test_rpcurl.py
@@ -1,14 +1,16 @@
 """Check that we correctly parse and serialize our URL format."""
+import re
+
 import pytest
 
 from shv import RpcLoginType, RpcProtocol, RpcUrl
 
 DATA = [
-    ("localsocket:/dev/null", RpcUrl("/dev/null", protocol=RpcProtocol.LOCAL_SOCKET)),
+    ("unix:/dev/null", RpcUrl("/dev/null", protocol=RpcProtocol.UNIX)),
     ("tcp://test@localhost:4242", RpcUrl("localhost", username="test", port=4242)),
     (
-        "udp://test@localhost:4242",
-        RpcUrl("localhost", protocol=RpcProtocol.UDP, username="test", port=4242),
+        "ssl://test@localhost:4242",
+        RpcUrl("localhost", protocol=RpcProtocol.SSL, username="test", port=4242),
     ),
     (
         "tcp://localhost:4242?devid=foo&devmount=/dev/null",
@@ -37,21 +39,35 @@ DATA = [
         ),
     ),
     ("tcp://[::]:4242", RpcUrl("::", port=4242)),
+    ("serial:/dev/ttyX", RpcUrl("/dev/ttyX", protocol=RpcProtocol.SERIAL)),
+    (
+        "serial:/dev/ttyX?baudrate=1152000",
+        RpcUrl("/dev/ttyX", protocol=RpcProtocol.SERIAL, baudrate=1152000),
+    ),
 ]
+
+
+@pytest.mark.parametrize("url,rpcurl", DATA)
+def test_to_url(url, rpcurl):
+    """Check that we correctly parse these URLs."""
+    assert rpcurl.to_url() == url
 
 
 @pytest.mark.parametrize(
     "url,rpcurl",
     DATA
     + [
-        ("", RpcUrl("", protocol=RpcProtocol.LOCAL_SOCKET)),
-        ("/dev/null", RpcUrl("/dev/null", protocol=RpcProtocol.LOCAL_SOCKET)),
-        ("//dev/null", RpcUrl("/dev/null", protocol=RpcProtocol.LOCAL_SOCKET)),
+        ("", RpcUrl("", protocol=RpcProtocol.UNIX)),
+        ("/dev/null", RpcUrl("/dev/null", protocol=RpcProtocol.UNIX)),
+        ("//dev/null", RpcUrl("/dev/null", protocol=RpcProtocol.UNIX)),
         (
-            "localsocket://dev/null",
-            RpcUrl("/dev/null", protocol=RpcProtocol.LOCAL_SOCKET),
+            "unix://dev/null",
+            RpcUrl("/dev/null", protocol=RpcProtocol.UNIX),
         ),
-        ("tcp://localhost", RpcUrl("localhost")),
+        ("tcp://localhost", RpcUrl("localhost", port=3755, protocol=RpcProtocol.TCP)),
+        ("tcps://localhost", RpcUrl("localhost", port=3765, protocol=RpcProtocol.TCPS)),
+        ("ssl://localhost", RpcUrl("localhost", port=3756, protocol=RpcProtocol.SSL)),
+        ("ssls://localhost", RpcUrl("localhost", port=3766, protocol=RpcProtocol.SSLS)),
         ("tcp://test@localhost:4242", RpcUrl("localhost", username="test", port=4242)),
         ("tcp://localhost?devid=foo", RpcUrl("localhost", device_id="foo")),
     ],
@@ -61,17 +77,17 @@ def test_parse(url, rpcurl):
     assert RpcUrl.parse(url) == rpcurl
 
 
-def test_invalid():
-    with pytest.raises(ValueError):
-        RpcUrl.parse("foo://some")
-
-
-def test_invalid_query():
-    with pytest.raises(ValueError):
-        RpcUrl.parse("tcp:localhost?missing=some")
-
-
-@pytest.mark.parametrize("url,rpcurl", DATA)
-def test_to_url(url, rpcurl):
-    """Check that we correctly parse these URLs."""
-    assert rpcurl.to_url() == url
+@pytest.mark.parametrize(
+    "url,msg",
+    (
+        ("tcp://localhost?missing=some", "Unsupported URL queries: missing"),
+        (
+            "tcp://localhost?missing=some&other=foo",
+            "Unsupported URL queries: missing, other",
+        ),
+        ("tcp:///dev/null", "Path is not supported for tcp: /dev/null"),
+    ),
+)
+def test_invalid(url, msg):
+    with pytest.raises(ValueError, match=re.escape(msg)):
+        RpcUrl.parse(url)